Move it back to the shared route + squash warning
